Create pom.xml to deploy webapp of Solr 5.0+
Starting from Solr ver. 5.0, its webapp is no longer deployed because
Solr is now intended to be used as a standalone service.

This repo contains the pom.xml file to deploy the webapp as a WAR file
(this archive does not contain all the dependencies that are usually
found in WEB-INF/lib).